generate/requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
lark>=1.1.0
tiktoken>=0.5.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
numpy>=1.24.0
